Fixed centring of slideshow in bounding box.
This fix amounts to hardcoding the width of the slideshow wrapper to the
same width as the surrounding element, as defined in main.css.
The proper fix will be to find a function to replace the previous one.
Previous function would get the window width programmatically, the new one
would need to find the width of the containing block or display element.