Removed slideshow size dependency between gallery page and css.
The hardcoded size in the javascript has been replaced by a function to
get the size of the first conataining <section></section> block.
This may not be ideal in future incarnations of the javascript but
at leasat the dependency is now contained within the gallery page.

There were comments to help track between the points of dependency, which
became redundant with this fix and were removed.